test: Updated Test Cases
